initial commit
Medical Center Portal Management System is developed using PHP, CSS, Bootstrap, and JavaScript. Talking about the project, it has all the essential features required for the user to view the hospital or pharmacy location and the price of a particular medicine or medical product by searching medicine or medical product name.